+++ lib/src/widgets/internal/controls/controls.dart
@@ -1,0 +1,2 @@
+export 'flutter_deck_controls.dart';
+export 'flutter_deck_controls_notifier.dart';